requirements.txt
PyQt5~=5.15.6

Pillow~=8.4.0

folium~=0.12.1.post1
